Trinity.Core: merge InternalCalls and TrinityC. Drop unused code.
