--- src/Trinity.Core/Trinity/Runtime/Trinity.C.cs
@@ -1,34 +1,0 @@
-// Graph Engine
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE.md file in the project root for full license information.
-//
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Security;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Trinity.Core.Lib
-{
-    class CTrinityC
-    {
-#if !CORECLR
-        [SecurityCritical]
-        [MethodImpl(MethodImplOptions.InternalCall)]
-#else
-        [DllImport(TrinityC.AssemblyName)]
-#endif
-        public static extern void TestEntry();
-
-#if !CORECLR
-        [SecurityCritical]
-        [MethodImpl(MethodImplOptions.InternalCall)]
-#else
-        [DllImport(TrinityC.AssemblyName)]
-#endif
-        internal static extern int C_GetLastError();
-    }
-}
